Fix: core dump caused by writing to the allnulls column in gtest
In the gtest setup test, the schema defines three fields that
do not allow null values. When arrow actually writing data to
these three columns, `arrow::MakeArrayOfNull` is used to construct
the column data.

For the types `arrow::int64` and `arrow::utf8`, this does not cause
a core dump. However, for the type `arrow::fixed_size_binary`, this
approach causes arrow to produce a core dump.

The relevant arrow logic is located in `column_writer.cc:L1782(WriteArrowSerialize)`:

```
Status WriteArrowSerialize(...)
{
	...
	bool no_nulls = // <---------- `no_nulls` always be true
      writer->descr()->schema_node()->is_required() || (array.null_count() == 0);

    if (!maybe_parent_nulls && no_nulls) {
      PARQUET_CATCH_NOT_OK(writer->WriteBatch(num_levels, def_levels, rep_levels, buffer));
  	} else {
      PARQUET_CATCH_NOT_OK(writer->WriteBatchSpaced(num_levels, def_levels, rep_levels,
                                                  array.null_bitmap_data(),
                                                  array.offset(), buffer));
	}
}
```

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>