enhance: refine the implemention of properties
Current commit reimplemented the properties code, integrating the
properties from the FFI (C interface) with those from C++.

The old version of the C++ properties had several issues:
- Each GetValue required dynamic casting of strings, which incurred
  performance overhead.
- It did not adequately define the type and range of properties, nor
  could it validate whether the value (string) was standardized in
  advance.
- The name of KEY was not defined.

The current changes address the three issues mentioned above. The
properties are now divided into two types: predefined properties
and non-predefined properties.

For predefined properties, the value (std::string) can be converted
to a specific type in advance (currently held via variant), and the
type of the property as well as the validity of the parameter range
can be verified early. For non-predefined properties, they are
uniformly processed as the `std::string` type.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>