enhance: replace the Status/Result with arrow defined one (#272)
In the milvus-storage, in a very early version, a `Status/Result` return
value mechanism was defined (the class names are the same as the classes
in Arrow), which partially encapsulated the Status/Result defined in
Arrow.

However, in the current version of the milvus-storage, many places mix
used of `arrow::Status/arrow::Result` and the milvus-storage defined
one. This reduces code readability. Current commit removed the
milvus-storage's defined one and replaces them with Arrow
implementation.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>