fix: ChunkReader can't pass the error when reader init failed (#276)
reproduce case:

1. create a reader(top-level) with a wrong manifest(invalid paths)
2. call `reader->get_record_batch_reader` will got a empty
`ArrowArrayStream`

This issue occurs in PackedRecordBatchReader::advanceBuffer. Due to the
lazy loading of the chunk reader, when methods of the chunk reader are
called in advanceBuffer, it mistakenly considers the current chunk
reader to be empty (which is allowed).

The lazy loading of the chunk reader causes the caller to be unable to
distinguish whether the current chunk reader is inherently empty or if
it is due to some other error. Therefore, the current change makes the
chunk reader no longer lazy-loaded and instead switches to using the
open method.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>