enhance: allow FFI function writer_new pass the manifest
The FFI function `writer_new` is used to create a storage writer.
After calling `writer_close`, a manifest will be generated and
returned to the caller.

If the caller repeatedly calls `writer_new` + `writer_close`,
multiple manifests will be generated. Therefore, in the previous
commit, the function `manifest_combine` was added which used to
combine multiple manifests. However, The function `manifest_combine`
cannot combine the manifests if their column groups differ.

To prevent the generation of multiple manifests with different
column groups under the same policy, `writer_new` now allows an
old manifest to be passed in. In practice, only the size-based
writer policy leads to this situation. Therefore, if the current
writer policy is size-based, the new writer will use the passed-in
old manifest to fix the column groups upon creation.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>