enhance: Introducing LRUCache to cache the filesystem (#284)
In the milvus-storage, an LRUCache (template class) has been introduced,
and the filesystem is cached using the filesystem config as the key.

Since the time required to create and destroy a filesystem is too long,
caching the filesystem is necessary. Since the filesystems used by
Parquet and Vortex are different (and there is very hard to create an
adapter between two of format), the cache objects used for different
entities in `format.cpp` are also different.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>