enhance: Use column groups to describe data files instead manifest.
The version in the `manifest` is not reflected in the
FFI interface. The current commit removes the `version`
field from the `manifest` and renames it to `column_groups`.
In fact, even before its removal, the `manifest` described
the data part of column groups. Throughout multiple version
iterations in storage, the `version` field still unused.
Also we highly likely won't use this field in the feature.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>