enhance: Introduce transaction in storage for FFI transaction API
In the previous commit (6d1f228), the version was removed
from column_groups, and it is no longer used as a manifest.

The current commit introduces the concept of a `transaction`.
This transaction is similar to the transaction in `Iceberg`,
but it only guarantees atomic reads and writes for a single path.
In Iceberg, the top-level implementation is the `dataset/table`.
However, since the `metadata` in the current Milvus has not been
offloaded, the storage layer can only handle the transaction
logic. This also results in some interfaces that should belong
to the `dataset/table` layer being placed at the transaction
layer. In the future, a new set of `table/catalog/transaction`
interfaces may be re-implemented, but for now, the current
interface is adequate.