enhance: Introduce new FFI API for caching layer
For the caching layer in Milvus, it requires the storage to precompute
chunks and provide corresponding metadata (including the number of rows
and memory size). Therefore, several new interfaces have been introduced
in the storage to address this requirement.

- get_column_group_infos: Used to parse the information contained in the
  column groups within column_groups.
- get_number_of_chunks: Retrieves the number of chunks in the current
  column group.
- get_chunk_metadatas: Retrieves the metadata of all chunks in the current
  column group.

After this commit , the `ChunkReader` will become a reader generated based
on "fixed" chunk divisions, while the `RecordBatchReader` will still as the
"streaming" reader. Note that when using the `ChunkReader`, the `Properties`
used to create the current chunk reader should not be altered to prevent
changes in the number of chunks. In contrast, the `RecordBatchReader` can
still use different properties to constrain the size of the returned record
batches.