fix: vector out of range in ChunkReaderImpl
There are two issues in ChunkReader.

1. It assumes that the number of `RecordBatches` returned by
`chunk_reader->get_chunks` equals the number of input chunk ids,
but in fact, the number of RecordBatches returned by get_chunks
can be less than or equal to the number of input chunk ids. This
is because after merging IOs, it might return a single RecordBatch
(both parquet and vortex).

2. For a single `chunk_reader`, concurrency cannot be performed, at
least for parquet(But vortex can). For the same file reader,
multi-threading will cause many issues.

The current commit disables the concurrency logic and re-write the
logical of `chunk_reader->get_chunks`. After get the result of get_chunks
it will call `Slice()` to make the number correspond to the number
of input chunk ids.

And there might be a complete concurrency solution that will default to
letting the caller manage concurrency resources in the future.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>