enhance: allow vortex reader use the logical size to split chunks (#298)
This commit adds support for the vortex chunk reader to split chunks 
using the logical chunk size (PROPERTY_READER_VORTEX_CHUNK_ROWS). 
Note that the final chunk size may not be equal to the logical size. 
because that vortex splits chunks according to the layout size by 
default, so when the logical chunk size exceeds the layout size,
the final chunk size becomes the layout size.

Additionally, this PR fixes an issue where `arrow::utf8` is read 
as `arrow::utf8view`. The vortex performs canonical encoding 
operations and converts the arrow schema according to the rules
`in vortex-dtype/src/arrow.rs` by default. Therefore, in the 
format reader, it is necessary to pass an `arrow::schema` to
ensure read type are consistent with write type.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>