enhance: Add start_index/end_index fields to column groups (#313)
The current changes are made to support external tables in stroage.
Since individual files of external tables can be relatively large, it is
necessary to perform "logical range" partitioning within the external
table files. To support it, the column groups have been augmented with
start_index and end_index fields.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>