enhance: Chunk reader support start/end indices.
In a previous commit (bfbe28c), the column group was modified
to support reading with start/end indices. The current commit
enables the chunk reader support reading with start/end indices.

`ColumnGroupReaderImpl` is specifically designed to handle the
relationship between `ChunkInfo()` and `RowGroupInfo()`. This
allows most of the logic to be reused, making the implementation
of the format reader simpler. However, vortex may experience
read amplification similar to parquet, which can be avoided.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>