enhance: add format reader interface to decouple chunk reader logic (#336)
In previous versions, a format-related reader was required to implement
the chunk reader interface. This interface needed to produce a set of
format readers and manage the reading in the chunks.

In practice, the two existing readers in storage (parquet and vortex)
both had significant redundant logic in chunk management. Therefore,
current commit extracts the common chunkreader logic and provides a new
format reader interface, allowing each format to focus solely on reading
individual files.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>