enhance: support format reader clone for multi-thread
This commit adds support for passing a thread pool when creating
the FFI reader, or creating a default thread pool if none is
provided. The thread count can be specified using the porperty
which named `PROPERTY_THREAD_POOL_SIZE`. Additionally, both
The function `get_chunks` and `take`  now support multithreading.

Since not all format readers are thread‑safe, a clone interface
is provided to ensure that format readers are not shared across
threads. Cloning a reader does not require reading the footer
(or other metadatas) from the S3 file. For readers that are
inherently thread‑safe, the original `shared_ptr` can be passed
directly; for non‑thread‑safe readers, a new reader can be
created by reading the same footer (or other metadata) without
needing to read the S3 file twice.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>