enhance: Allow columngroups export/import and ajust the external table interface (#335)
This commit defines the composition of column groups in the FFI and
provides interfaces to export/import the structure of column groups.
Also adjusts the external table interface:

- `column_groups_export`: Exports the column groups object from C++ into
a C struct.
- `column_groups_import`: Imports the column groups C struct into a C++
object.
- `exttable_explore`: Lists the input explore_dir and generates column
groups, outputting the number of files under explore_dir and the
location of the temporary column group files created.
- `exttable_read_column_groups`: Reads column groups by file path and
outputs the column groups as a C struct.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>