fix: The filesystem config root_path has no effect.
The `root_path` parameter is designed for the `LocalFileSystem`,
allowing it to read and write under a unified directory path.
However, when this parameter was initially implemented, it only
performed directory creation and did not actually make any
operations performed through the `LocalFileSystem` under the
specified `root_path`.

The current change wraps the `LocalFileSystem` with the
`SubTreeFileSystem`, ensuring that `root_path` takes effect.
If a relative path is passed as `root_path`, the storage will
convert it to an absolute path (relative to the process’s
working directory, caller can use `chdir` change it).

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>