enhance: move multipart-upload-size to filesystem config.
In the parquet writer, there has always been the practice of forcibly casting `arrow::fs::FileSystem` to the `class MultiPartUploadS3FS` and then calling methods not provided by `arrow::fs::FileSystem`. This breaks the encapsulation of the Arrow file system.

The current commit removes the forced cast of file system and allows users to decide the multipart upload size when creating the `arrow::fs::FileSystem`.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>